add finnaly on try/catch to redirect

+++ app/Http/Controllers/Admin/OrderController.php
@@ -43,10 +43,14 @@ class OrderController extends Controller
         try {
             $order = Order::findOrFail($id);
             $orderDetail = OrderDetail::query()->where('order_id', $id)->delete();
-            $order->delete();
+            if ($orderDetail) {
+                $order->delete();
+            }
             session(['message' => __('orders.admin.list.deleted')]);
         } catch (ModelNotFoundException $e) {
             session(['message' => __('orders.admin.list.id_not_found')]);
+        } finally {
+            return redirect()->route('admin.orders.index');
         }
     }
 }
